Modified Book controller

+++ src/main/java/com/suhail/demo/Book.java
@@ -1,17 +1,17 @@
 package com.suhail.demo;
 
 public class Book {
-	int id;
+	long id;
 	String name;
 	String author;
 	
-	public Book(int id, String name, String author) {
+	public Book(long id, String name, String author) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.author = author;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 	public String getName() {
